+++ pkg/addons/addon_config.go
@@ -139,7 +139,9 @@ func extractRef(rNodes []*yaml.RNode, lookupPath []string) string {
 		if ref == nil || err != nil {
 			continue
 		}
-		refStr, _ := ref.String()
+
+		refYNode := ref.YNode() // Convert to go yaml.node to extract values without trailing newlines
+		refStr := refYNode.Value
 		log.Debugf("found ref: %s", refStr)
 		return refStr
 	}