+++ test/gen_integration.sh
@@ -1,9 +1,14 @@
+export WORKFLOWS_PATH=.github/workflows
+
 # remove previous tests
 echo "Removing previous integration configs"
 rm -rf ./integration/*
 echo "Removing previous integration workflows"
-rm ../.github/workflows/integration-linux.yml
-rm ../.github/workflows/integration-windows.yml
+rm ../$WORKFLOWS_PATH/integration-linux.yml
+rm ../$WORKFLOWS_PATH/integration-windows.yml
+
+# create temp files for keeping track off workflow jobs to build job-dependency graph
+# this is used to populated the needs: field of the required final workflow jobs
 helm_workflow_names_file=./temp/helm_workflow_names.txt
 rm $helm_workflow_names_file
 helm_win_workflow_names_file=./temp/helm_win_workflow_names.txt
@@ -14,23 +19,22 @@ kustomize_win_workflow_names_file=./temp/kustomize_win_workflow_names.txt
 rm $kustomize_win_workflow_names_file
 manifest_workflow_names_file=./temp/manifest_workflow_names.txt
 rm $manifest_workflow_names_file
+mkdir -p ./temp
 
 # add build to workflow
 echo "# this file is generated using gen_integration.sh
 name: draft Linux Integrations
-
 on:
   pull_request:
     branches: [ main ]
   workflow_dispatch:
-
 jobs:
   build:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
       - name: Set up Go
-        uses: actions/setup-go@v2
+        uses: actions/setup-go@v4
         with:
           go-version: 1.18.2
       - name: make
@@ -44,15 +48,13 @@ jobs:
         with:
           name: draft-binary
           path: ./draft
-          if-no-files-found: error" > ../.github/workflows/integration-linux.yml
+          if-no-files-found: error" > ../$WORKFLOWS_PATH/integration-linux.yml
 
 echo "name: draft Windows Integrations
-
 on:
   pull_request_review:
     types: [submitted]
   workflow_dispatch:
-
 jobs:
   build:
     runs-on: windows-latest
@@ -88,7 +90,7 @@ jobs:
         with:
           name: check_windows_addon_kustomize
           path: ./test/check_windows_addon_kustomize.ps1
-          if-no-files-found: error" > ../.github/workflows/integration-windows.yml
+          if-no-files-found: error" > ../$WORKFLOWS_PATH/integration-windows.yml
 
 
 # read config and add integration test for each language
@@ -102,8 +104,11 @@ do
     port=$(echo $test | jq '.port' -r)
     serviceport=$(echo $test | jq '.serviceport' -r)
     repo=$(echo $test | jq '.repo' -r)
+
+    imagename="host.minikube.internal:5001/testapp"
     # addon integration testing vars
     ingress_test_args="-a webapp_routing --variable ingress-tls-cert-keyvault-uri=test.cert.keyvault.uri --variable ingress-use-osm-mtls=true --variable ingress-host=host1"
+    create_config_args="--variable PORT=8080 --variable APPNAME=testingCreateCommand --variable VERSION=1.11 --variable BUILDERVERSION=1.11 --variable SERVICEPORT=8080 --variable NAMESPACE=testNamespace --variable IMAGENAME=testImage --variable IMAGETAG=latest"
     echo "Adding $lang with port $port"
 
     mkdir ./integration/$lang
@@ -119,6 +124,8 @@ deployVariables:
     value: \"$serviceport\"
   - name: \"APPNAME\"
     value: \"testapp\"
+  - name: \"IMAGENAME\"
+    value: \"$imagename\"
 languageVariables:
   - name: \"VERSION\"
     value: \"$version\"
@@ -138,6 +145,8 @@ deployVariables:
     value: \"$serviceport\"
   - name: \"APPNAME\"
     value: \"testapp\"
+  - name: \"IMAGENAME\"
+    value: \"$imagename\"
 languageVariables:
   - name: \"VERSION\"
     value: \"$version\"
@@ -146,7 +155,7 @@ languageVariables:
   - name: \"PORT\"
     value: \"$port\"" > ./integration/$lang/kustomize.yaml
 
-    # create kustomize.yaml
+    # create manifest.yaml
     echo "$note
 deployType: \"manifests\"
 languageType: \"$lang\"
@@ -157,6 +166,8 @@ deployVariables:
     value: \"$serviceport\"
   - name: \"APPNAME\"
     value: \"testapp\"
+  - name: \"IMAGENAME\"
+    value: \"$imagename\"
 languageVariables:
   - name: \"VERSION\"
     value: \"$version\"
@@ -183,12 +194,22 @@ languageVariables:
           with:
             repository: $repo
             path: ./langtest
-        - name: Execute Dry Run
+        - name: Execute Dry Run with config file
           run: |
             mkdir -p test/temp
             ./draft --dry-run --dry-run-file test/temp/dry-run.json \
             create -c ./test/integration/$lang/helm.yaml \
             -d ./langtest/ --skip-file-detection
+        - name: Validate JSON
+          run: |
+            npm install -g ajv-cli@5.0.0
+            ajv validate -s test/dry_run_schema.json -d test/temp/dry-run.json
+        - name: Execute Dry Run with variables passed through flag 
+          run: |
+            mkdir -p test/temp
+            ./draft --dry-run --dry-run-file test/temp/dry-run.json \
+            create -d ./langtest/ -l $lang --skip-file-detection --deploy-type helm \
+            $create_config_args
         - name: Validate JSON
           run: |
             npm install -g ajv-cli@5.0.0
@@ -199,7 +220,7 @@ languageVariables:
       registry:
         image: registry:2
         ports:
-          - 5000:5000
+          - 5001:5000
     needs: $lang-helm-dry-run
     steps:
       - uses: actions/checkout@v3
@@ -214,21 +235,14 @@ languageVariables:
           path: ./langtest
       - run: rm -rf ./langtest/manifests && rm -f ./langtest/Dockerfile ./langtest/.dockerignore
       - run: ./draft -v create -c ./test/integration/$lang/helm.yaml -d ./langtest/
-      - run: ./draft -b main -v generate-workflow -d ./langtest/ -c someAksCluster -r someRegistry -g someResourceGroup --container-name someContainer
-      - run: ./draft -v update -d ./langtest/ $ingress_test_args
       - name: start minikube
         id: minikube
         uses: medyagh/setup-minikube@master
-      - name: Build image
-        run: |
-          export SHELL=/bin/bash
-          eval \$(minikube -p minikube docker-env)
-          docker build -f ./langtest/Dockerfile -t testapp ./langtest/
-          echo -n "verifying images:"
-          docker images
+        with:
+          insecure-registry: 'host.minikube.internal:5001,10.0.0.0/24'
       # Runs Helm to create manifest files
       - name: Bake deployment
-        uses: azure/k8s-bake@v2.1
+        uses: azure/k8s-bake@v2.2
         with:
           renderEngine: 'helm'
           helmChart: ./langtest/charts
@@ -236,21 +250,81 @@ languageVariables:
           overrides: |
             replicas:2
           helm-version: 'latest'
+          releaseName: 'test-release'
         id: bake
+      - name: Build and Push image
+        continue-on-error: true
+        run: |
+          export SHELL=/bin/bash
+          eval \$(minikube -p minikube docker-env)
+          docker build -f ./langtest/Dockerfile -t testapp ./langtest/
+          docker tag testapp $imagename
+          echo -n \"verifying images:\"
+          docker images
+          docker push $imagename
+          echo 'Curling host.minikube.internal test app images from minikube'
+          minikube ssh \"curl http://host.minikube.internal:5001/v2/testapp/tags/list\"
       # Deploys application based on manifest files from previous step
       - name: Deploy application
-        uses: Azure/k8s-deploy@v3.0
+        uses: Azure/k8s-deploy@v4.0
         continue-on-error: true
         id: deploy
         with:
           action: deploy
           manifests: \${{ steps.bake.outputs.manifestsBundle }}
+          images: |
+            $imagename
+      - name: Wait for rollout
+        continue-on-error: true
+        id: rollout
+        run: |
+          kubectl rollout status deployment/test-release-testapp --timeout=2m
+      - name: Print K8s Objects
+        run: |
+          kubectl get po -o json
+          kubectl get svc -o json
+          kubectl get deploy -o json
+      - name: Curl Endpoint
+        run: |
+          kubectl get svc
+          echo 'Starting minikube tunnel'
+          minikube tunnel  > /dev/null 2>&1 & tunnelPID=\$!
+          sleep 120
+          kubectl get svc
+          SERVICEIP=\$(kubectl get svc -o jsonpath={'.items[1].status.loadBalancer.ingress[0].ip'})
+          echo \"SERVICEIP: \$SERVICEIP\"
+          echo 'Curling service IP'
+          curl -m 3 \$SERVICEIP:$serviceport
+          kill \$tunnelPID
+      - run: |
+          ./draft -b main -v generate-workflow -d ./langtest/ -c someAksCluster -r someRegistry -g someResourceGroup --container-name someContainer --deploy-type helm --build-context-path .
+          pwd
+      # Validate generated workflow yaml
+      - name: Install action-validator with asdf
+        uses: asdf-vm/actions/install@v1
+        with:
+          tool_versions: |
+            action-validator 0.1.2
+      - name: Lint Actions
+        run: |
+          find $WORKFLOWS_PATH -type f \( -iname \*.yaml -o -iname \*.yml \) \
+            | xargs -I {} action-validator --verbose {}
+      - name: Execute dry run for update command
+        run: |
+          mkdir -p test/temp
+          pwd
+          ./draft --dry-run --dry-run-file test/temp/update_dry_run.json update -d ./langtest/ $ingress_test_args  
+      - name: Validate JSON
+        run: |
+          npm install -g ajv-cli@5.0.0
+          ajv validate -s test/update_dry_run_schema.json -d test/temp/update_dry_run.json
+      - run: ./draft -v update -d ./langtest/ $ingress_test_args
       - name: Check default namespace
         if: steps.deploy.outcome != 'success'
         run: kubectl get po
       - name: Fail if any error
         if: steps.deploy.outcome != 'success'
-        run: exit 6" >> ../.github/workflows/integration-linux.yml
+        run: exit 6" >> ../$WORKFLOWS_PATH/integration-linux.yml
 
     # create kustomize workflow
     kustomize_create_update_job_name=$lang-kustomize-create-update
@@ -270,12 +344,22 @@ languageVariables:
         with:
           repository: $repo
           path: ./langtest
-      - name: Execute Dry Run
+      - name: Execute Dry Run with config file
         run: |
           mkdir -p test/temp
           ./draft --dry-run --dry-run-file test/temp/dry-run.json \
           create -c ./test/integration/$lang/kustomize.yaml \
           -d ./langtest/ --skip-file-detection
+      - name: Validate JSON
+        run: |
+          npm install -g ajv-cli@5.0.0
+          ajv validate -s test/dry_run_schema.json -d test/temp/dry-run.json
+      - name: Execute Dry Run with variables passed through flag 
+        run: |
+          mkdir -p test/temp
+          ./draft --dry-run --dry-run-file test/temp/dry-run.json \
+          create -d ./langtest/ -l $lang --skip-file-detection --deploy-type kustomize \
+          $create_config_args
       - name: Validate JSON
         run: |
           npm install -g ajv-cli@5.0.0
@@ -286,7 +370,7 @@ languageVariables:
       registry:
         image: registry:2
         ports:
-          - 5000:5000
+          - 5001:5000
     needs: $lang-kustomize-dry-run
     steps:
       - uses: actions/checkout@v3
@@ -301,41 +385,87 @@ languageVariables:
           path: ./langtest
       - run: rm -rf ./langtest/manifests && rm -f ./langtest/Dockerfile ./langtest/.dockerignore
       - run: ./draft -v create -c ./test/integration/$lang/kustomize.yaml -d ./langtest/
-      - run: ./draft -v generate-workflow -b main -d ./langtest/ -c someAksCluster -r someRegistry -g someResourceGroup --container-name someContainer
-      - run: ./draft -v update -d ./langtest/ $ingress_test_args
       - name: start minikube
         id: minikube
         uses: medyagh/setup-minikube@master
+        with:
+          insecure-registry: 'host.minikube.internal:5001,10.0.0.0/24'
       - name: Bake deployment
-        uses: azure/k8s-bake@v2.1
+        uses: azure/k8s-bake@v2.4
+        id: bake
         with:
           renderEngine: 'kustomize'
           kustomizationPath: ./langtest/base
           kubectl-version: 'latest'
-        id: bake
-      - name: Build image
+      - name: Build and Push Image
+        continue-on-error: true
         run: |
-          export SHELL=/bin/bash
-          eval \$(minikube -p minikube docker-env)
-          docker build -f ./langtest/Dockerfile -t testapp:curr ./langtest/
-          echo -n "verifying images:"
+          eval \$(minikube docker-env)
+          docker build -f ./langtest/Dockerfile -t testapp ./langtest/
+          docker tag testapp $imagename
+          echo -n \"verifying images:\"
           docker images
+          docker push $imagename
+          echo 'Curling host.minikube.internal test app images from minikube'
+          minikube ssh \"curl http://host.minikube.internal:5001/v2/testapp/tags/list\"
       # Deploys application based on manifest files from previous step
       - name: Deploy application
-        uses: Azure/k8s-deploy@v3.0
+        uses: Azure/k8s-deploy@v4.0
         continue-on-error: true
         id: deploy
         with:
           action: deploy
           manifests: \${{ steps.bake.outputs.manifestsBundle }}
           images: |
-            testapp:curr
+            $imagename
+      - name: Wait for rollout
+        continue-on-error: true
+        id: rollout
+        run: |
+          kubectl rollout status deployment/testapp --timeout=2m
+      - name: Print K8s Objects
+        run: |
+          kubectl get po -o json
+          kubectl get svc -o json
+          kubectl get deploy -o json
+      - name: Curl Endpoint
+        run: |
+          kubectl get svc
+          echo 'Starting minikube tunnel'
+          minikube tunnel  > /dev/null 2>&1 & tunnelPID=\$!
+          sleep 120
+          kubectl get svc
+          SERVICEIP=\$(kubectl get svc -o jsonpath={'.items[1].status.loadBalancer.ingress[0].ip'})
+          echo \"SERVICEIP: \$SERVICEIP\"
+          echo 'Curling service IP'
+          curl -m 3 \$SERVICEIP:$serviceport
+          kill \$tunnelPID
+      - run: ./draft -v generate-workflow -b main -d ./langtest/ -c someAksCluster -r someRegistry -g someResourceGroup --container-name someContainer --deploy-type kustomize --build-context-path .
+      # Validate generated workflow yaml
+      - name: Install action-validator with asdf
+        uses: asdf-vm/actions/install@v1
+        with:
+          tool_versions: |
+            action-validator 0.1.2
+      - name: Lint Actions
+        run: |
+          find $WORKFLOWS_PATH -type f \( -iname \*.yaml -o -iname \*.yml \) \
+            | xargs -I {} action-validator --verbose {}
+      - name: Execute dry run for update command
+        run: |
+          mkdir -p test/temp
+          ./draft --dry-run --dry-run-file test/temp/update_dry_run.json update -d ./langtest/ $ingress_test_args  
+      - name: Validate JSON
+        run: |
+          npm install -g ajv-cli@5.0.0
+          ajv validate -s test/update_dry_run_schema.json -d test/temp/update_dry_run.json
+      - run: ./draft -v update -d ./langtest/ $ingress_test_args
       - name: Check default namespace
         if: steps.deploy.outcome != 'success'
         run: kubectl get po
       - name: Fail if any error
         if: steps.deploy.outcome != 'success'
-        run: exit 6" >> ../.github/workflows/integration-linux.yml
+        run: exit 6" >> ../$WORKFLOWS_PATH/integration-linux.yml
 
   # create manifests workflow
     manifest_update_job_name=$lang-manifest-update
@@ -355,12 +485,22 @@ languageVariables:
           with:
             repository: $repo
             path: ./langtest
-        - name: Execute Dry Run
+        - name: Execute Dry Run with config file
           run: |
             mkdir -p test/temp
             ./draft --dry-run --dry-run-file test/temp/dry-run.json \
             create -c ./test/integration/$lang/manifest.yaml \
             -d ./langtest/ --skip-file-detection
+        - name: Validate JSON
+          run: |
+            npm install -g ajv-cli@5.0.0
+            ajv validate -s test/dry_run_schema.json -d test/temp/dry-run.json
+        - name: Execute Dry Run with variables passed through flag 
+          run: |
+            mkdir -p test/temp
+            ./draft --dry-run --dry-run-file test/temp/dry-run.json \
+            create -d ./langtest/ -l $lang --skip-file-detection --deploy-type manifests \
+            $create_config_args
         - name: Validate JSON
           run: |
             npm install -g ajv-cli@5.0.0
@@ -371,7 +511,7 @@ languageVariables:
       registry:
         image: registry:2
         ports:
-          - 5000:5000
+          - 5001:5000
     needs: $lang-manifest-dry-run
     steps:
       - uses: actions/checkout@v3
@@ -386,29 +526,72 @@ languageVariables:
           path: ./langtest
       - run: rm -rf ./langtest/manifests && rm -f ./langtest/Dockerfile ./langtest/.dockerignore
       - run: ./draft -v create -c ./test/integration/$lang/manifest.yaml -d ./langtest/
-      - run: ./draft -v generate-workflow -d ./langtest/ -b main -c someAksCluster -r someRegistry -g someResourceGroup --container-name someContainer
+      - name: print manifests
+        run: cat ./langtest/manifests/*
+      - name: Add docker.local host to /etc/hosts
+        run: |
+          sudo echo \"127.0.0.1 docker.local\" | sudo tee -a /etc/hosts
       - name: start minikube
         id: minikube
         uses: medyagh/setup-minikube@master
-      - name: Build image
+        with:
+          insecure-registry: 'host.minikube.internal:5001,10.0.0.0/24'
+      - name: Build and Push Image
+        continue-on-error: true
         run: |
-          export SHELL=/bin/bash
-          eval \$(minikube -p minikube docker-env)
+          eval \$(minikube docker-env)
           docker build -f ./langtest/Dockerfile -t testapp ./langtest/
-          echo -n "verifying images:"
+          docker tag testapp $imagename
+          echo -n \"verifying images:\"
           docker images
+          docker push $imagename
+          echo 'Curling host.minikube.internal test app images from minikube'
+          minikube ssh \"curl http://host.minikube.internal:5001/v2/testapp/tags/list\"
       # Deploys application based on manifest files from previous step
       - name: Deploy application
         run: kubectl apply -f ./langtest/manifests/
         continue-on-error: true
         id: deploy
-      - name: Check default namespace
-        if: steps.deploy.outcome != 'success'
-        run: kubectl get po
+      - name: Wait for rollout
+        continue-on-error: true
+        id: rollout
+        run: |
+          kubectl rollout status deployment/testapp --timeout=2m
+      - name: Print K8s Objects
+        run: |
+          kubectl get po -o json
+          kubectl get svc -o json
+          kubectl get deploy -o json
+      - name: Curl Endpoint
+        run: |
+          kubectl get svc
+          echo 'Starting minikube tunnel'
+          minikube tunnel  > /dev/null 2>&1 & tunnelPID=\$!
+          sleep 120
+          kubectl get svc
+          SERVICEIP=\$(kubectl get svc -o jsonpath={'.items[1].status.loadBalancer.ingress[0].ip'})
+          echo \"SERVICEIP: \$SERVICEIP\"
+          echo 'Curling service IP'
+          curl -m 3 \$SERVICEIP:$serviceport
+          kill \$tunnelPID
+      - run: ./draft -v generate-workflow -d ./langtest/ -b main -c someAksCluster -r localhost -g someResourceGroup --container-name testapp --deploy-type manifests --build-context-path .
+      # Validate generated workflow yaml
+      - name: Install action-validator with asdf
+        uses: asdf-vm/actions/install@v1
+        with:
+          tool_versions: |
+            action-validator 0.1.2
+      - name: Lint Actions
+        run: |
+          find $WORKFLOWS_PATH -type f \( -iname \*.yaml -o -iname \*.yml \) \
+            | xargs -I {} action-validator --verbose {}
       - uses: actions/upload-artifact@v3
         with:
           name: $lang-manifests-create
           path: ./langtest
+      - name: Fail if any error
+        if: steps.deploy.outcome != 'success' || steps.rollout.outcome != 'success'
+        run: exit 6
   $manifest_update_job_name:
     needs: $lang-manifests-create
     runs-on: ubuntu-latest
@@ -427,6 +610,14 @@ languageVariables:
         with:
           name: $lang-manifests-create
           path: ./langtest/
+      - name: Execute dry run for update command
+        run: |
+          mkdir -p test/temp
+          ./draft --dry-run --dry-run-file test/temp/update_dry_run.json update -d ./langtest/ $ingress_test_args  
+      - name: Validate JSON
+        run: |
+          npm install -g ajv-cli@5.0.0
+          ajv validate -s test/update_dry_run_schema.json -d test/temp/update_dry_run.json
       - run: ./draft -v update -d ./langtest/ $ingress_test_args
       - name: start minikube
         id: minikube
@@ -445,7 +636,10 @@ languageVariables:
         id: deploy
       - name: Check default namespace
         if: steps.deploy.outcome != 'success'
-        run: kubectl get po" >> ../.github/workflows/integration-linux.yml
+        run: kubectl get po
+      - name: Fail if any error
+        if: steps.deploy.outcome != 'success'
+        run: exit 6" >> ../$WORKFLOWS_PATH/integration-linux.yml
 
   helm_update_win_jobname=$lang-helm-update
   echo $helm_update_win_jobname >> $helm_win_workflow_names_file
@@ -497,7 +691,7 @@ languageVariables:
           name: check_windows_addon_helm
           path: ./langtest/
       - run: ./check_windows_addon_helm.ps1
-        working-directory: ./langtest/" >> ../.github/workflows/integration-windows.yml
+        working-directory: ./langtest/" >> ../$WORKFLOWS_PATH/integration-windows.yml
 
     # create kustomize workflow
     kustomize_win_workflow_name=$lang-kustomize-update
@@ -549,7 +743,7 @@ languageVariables:
           path: ./langtest/
       - run: ./check_windows_addon_kustomize.ps1
         working-directory: ./langtest/
-      " >> ../.github/workflows/integration-windows.yml
+      " >> ../$WORKFLOWS_PATH/integration-windows.yml
 done
 
 echo "
@@ -558,7 +752,7 @@ echo "
       needs: [ $( paste -sd ',' $helm_win_workflow_names_file) ]
       steps:
         - run: echo "helm integrations passed"
-" >> ../.github/workflows/integration-windows.yml
+" >> ../$WORKFLOWS_PATH/integration-windows.yml
 
 echo "
   kustomize-win-integrations-summary:
@@ -566,7 +760,7 @@ echo "
       needs: [ $( paste -sd ',' $kustomize_win_workflow_names_file) ]
       steps:
         - run: echo "kustomize integrations passed"
-" >> ../.github/workflows/integration-windows.yml
+" >> ../$WORKFLOWS_PATH/integration-windows.yml
 
 echo "
   helm-integrations-summary:
@@ -574,7 +768,7 @@ echo "
       needs: [ $( paste -sd ',' $helm_workflow_names_file) ]
       steps:
         - run: echo "helm integrations passed"
-" >> ../.github/workflows/integration-linux.yml
+" >> ../$WORKFLOWS_PATH/integration-linux.yml
 
 echo "
   kustomize-integrations-summary:
@@ -582,7 +776,7 @@ echo "
       needs: [ $( paste -sd ',' $kustomize_workflow_names_file) ]
       steps:
         - run: echo "kustomize integrations passed"
-" >> ../.github/workflows/integration-linux.yml
+" >> ../$WORKFLOWS_PATH/integration-linux.yml
 
 echo "
   manifest-integrations-summary:
@@ -590,4 +784,4 @@ echo "
       needs: [ $( paste -sd ',' $manifest_workflow_names_file) ]
       steps:
         - run: echo "manifest integrations passed"
-" >> ../.github/workflows/integration-linux.yml
+" >> ../$WORKFLOWS_PATH/integration-linux.yml